#1 Move announcement into separate file
